+++ scylla/tests/ccm-integration/ccm/node_config.rs
