+++ scylla/tests/ccm-integration/main.rs
@@ -1,6 +1,2 @@
-extern crate core;
-
-pub(crate) mod cluster;
-pub(crate) mod logged_cmd;
-pub(crate) mod node_config;
+pub(crate) mod ccm;
 mod test_example;
